requirements.txt
streamlit==1.55.0
python-dotenv==1.0.1
joblib>=1.3.0
langgraph==0.2.50
langchain==0.3.7
langchain-core==0.3.15
langchain-community==0.3.5
langchain-google-genai==2.0.5
pydantic==2.8.2
pydantic-settings>=2.7,<3
rouge-score==0.1.2
loguru==0.7.2
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
typing-extensions>=4.8.0
importlib-metadata>=6.8.0
pytest>=7.4.0
google-generativeai>=0.3.0
email-validator>=2.0.0
# python-magic>=0.4.27
requests>=2.31.0
